requirements.txt
Django
djangorestframework
django-filter
django-filters
pytest
django-pytest
factory-boy
django-cors-headers
django-webpack-loader